Create README.md
I am an energetic, straight forward and passionate
undergraduate with an outstanding academic performances in
IT industry. I am interested in cyber security, data science, software
engineering. I always like to take problems as challenges and solve
creatively and efficiently an individual or working as a team. With
this, I am looking forward to the opportunities which suit for
undergraduate level.